refactor: update shipping calculators to plugins
