+++ packages/melhor-envio/src/plugin.ts
@@ -1,0 +1,13 @@
+import { PluginCommonModule, VendurePlugin } from '@vendure/core';
+import { MelhorEnvioShippingCalculator } from './shipping-calculator';
+
+@VendurePlugin({
+  imports: [PluginCommonModule],
+  configuration: (config) => {
+    config.shippingOptions.shippingCalculators?.push(
+      MelhorEnvioShippingCalculator
+    );
+    return config;
+  }
+})
+export class MelhorEnvioPlugin {}
