+++ packages/ups-brazil/src/plugin.ts
@@ -1,0 +1,13 @@
+import { PluginCommonModule, VendurePlugin } from '@vendure/core';
+import { UPSBrazilShippingCalculator } from './shipping-calculator';
+
+@VendurePlugin({
+  imports: [PluginCommonModule],
+  configuration: (config) => {
+    config.shippingOptions.shippingCalculators?.push(
+      UPSBrazilShippingCalculator
+    );
+    return config;
+  }
+})
+export class UPSBrazilPlugin {}
